# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Upsert request including a cryptographic identity verification bundle generated by iOS GameKit's generateIdentityVerificationSignatureAsync.

 */
export interface SyncPlayerRequest {
  /** Team-scoped or game-scoped Game Center player ID */
  gameCenterId: string;
  displayName: string;
  /** Apple CDN URL for the public key (*.apple.com only) */
  publicKeyUrl: string;
  /** Base64-encoded RSA-SHA256 signature over playerID+bundleID+salt+timestamp */
  signature: string;
  /** Base64-encoded random salt */
  salt: string;
  /** Unix timestamp in milliseconds as returned by iOS */
  timestamp: number;
}

export interface PlayerProfile {
  id: number;
  gameCenterId: string;
  displayName: string;
  coins: number;
  wins: number;
  losses: number;
  winStreak: number;
  elo: number;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}