# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: players
    description: Player profile operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /players/sync:
    post:
      operationId: syncPlayer
      tags: [players]
      summary: Upsert player by verified Game Center ID
      description: >
        Creates or updates a player profile keyed by their Game Center ID.
        Requires a cryptographic identity verification bundle (publicKeyUrl,
        signature, salt, timestamp) generated by iOS GameKit — raw IDs are
        rejected to prevent spoofing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncPlayerRequest"
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Game Center identity verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /players/{gameCenterId}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Fetch player profile
      description: Returns the stored profile for a given verified Game Center player ID.
      parameters:
        - name: gameCenterId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerProfile"
        "404":
          description: Player not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    SyncPlayerRequest:
      type: object
      description: >
        Upsert request including a cryptographic identity verification bundle
        generated by iOS GameKit's generateIdentityVerificationSignatureAsync.
      properties:
        gameCenterId:
          type: string
          description: Team-scoped or game-scoped Game Center player ID
        displayName:
          type: string
        publicKeyUrl:
          type: string
          format: uri
          description: Apple CDN URL for the public key (*.apple.com only)
        signature:
          type: string
          description: Base64-encoded RSA-SHA256 signature over playerID+bundleID+salt+timestamp
        salt:
          type: string
          description: Base64-encoded random salt
        timestamp:
          type: integer
          description: Unix timestamp in milliseconds as returned by iOS
      required:
        - gameCenterId
        - displayName
        - publicKeyUrl
        - signature
        - salt
        - timestamp
    PlayerProfile:
      type: object
      properties:
        id:
          type: integer
        gameCenterId:
          type: string
        displayName:
          type: string
        coins:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winStreak:
          type: integer
        elo:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - gameCenterId
        - displayName
        - coins
        - wins
        - losses
        - winStreak
        - elo
        - createdAt
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error