# lib/api-zod/src/generated/types/syncPlayerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Upsert request including a cryptographic identity verification bundle generated by iOS GameKit's generateIdentityVerificationSignatureAsync.

 */
export interface SyncPlayerRequest {
  /** Team-scoped or game-scoped Game Center player ID */
  gameCenterId: string;
  displayName: string;
  /** Apple CDN URL for the public key (*.apple.com only) */
  publicKeyUrl: string;
  /** Base64-encoded RSA-SHA256 signature over playerID+bundleID+salt+timestamp */
  signature: string;
  /** Base64-encoded random salt */
  salt: string;
  /** Unix timestamp in milliseconds as returned by iOS */
  timestamp: number;
}